+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-2012-09-26  Michiel Roos <michiel@maxserv.nl>
+2018-12-24  Lex Frolenko
+- [TASK] TYPO3 9.5 compatibility
+- [BUGFIX] fixed pagetree parent for success page
+- [BUGFIX] fixed display hidden pages for success page
+
+2015-11-11  Michiel Roos <michiel@maxserv.com>
+- [TASK] TYPO3 7.6 compatibility
+
+2014-09-15  Michiel Roos <michiel@maxserv.com>
+- [TASK] TYPO3 6.2 compatibility
+
+2012-09-26  Michiel Roos <michiel@maxserv.com>
 - Refactor Language setup
 - Add ext_autoload.php
 - Move images into Resources folder